+++ sandbox/replication.go
@@ -19,8 +19,8 @@ import (
 	"fmt"
 	"os"
 	"path"
-	"time"
 	"strings"
+	"time"
 
 	"github.com/datacharmer/dbdeployer/common"
 	"github.com/datacharmer/dbdeployer/concurrent"
@@ -184,7 +184,7 @@ func CreateMasterSlaveReplication(sandboxDef SandboxDef, origin string, nodes in
 			autoPosOpt = "MASTER_AUTO_POSITION=1"
 		}
 		masterAutoPosition += ", " + autoPosOpt
-		logger.Printf("Adding %s to slaves setup\n", autoPosOpt)		
+		logger.Printf("Adding %s to slaves setup\n", autoPosOpt)
 	}
 	// 8.0.11
 	// isMinimumNativeAuthPlugin, err := common.GreaterOrEqualVersion(sandboxDef.Version, globals.MinimumNativeAuthPluginVersion)
@@ -557,6 +557,16 @@ func CreateReplicationSandbox(sdef SandboxDef, origin string, replData Replicati
 				common.IntSliceToDottedString(globals.MinimumNdbClusterVersion))
 		}
 		sdef.SandboxDir = path.Join(sdef.SandboxDir, defaults.Defaults().NdbPrefix+common.VersionToName(origin))
+	case globals.InnoDBClusterLabel:
+		isMinimumInnodbCluster, err := common.HasCapability(sdef.Flavor, common.InnoDBCluster, sdef.Version)
+		if err != nil {
+			return err
+		}
+		if !isMinimumInnodbCluster {
+			return fmt.Errorf(globals.ErrFeatureRequiresCapability, "InnoDB Cluster", common.MySQLFlavor,
+				common.IntSliceToDottedString(globals.MinimumInnoDBCluster))
+		}
+		sdef.SandboxDir = path.Join(sdef.SandboxDir, defaults.Defaults().InnoDBClusterPrefix+common.VersionToName(origin))
 	default:
 		return fmt.Errorf("unrecognized topology. Accepted: '%v'", globals.AllowedTopologies)
 	}
@@ -589,6 +599,8 @@ func CreateReplicationSandbox(sdef SandboxDef, origin string, replData Replicati
 		err = CreatePxcReplication(sdef, origin, replData.Nodes, replData.MasterIp)
 	case globals.NdbLabel:
 		err = CreateNdbReplication(sdef, origin, replData.Nodes, replData.NdbNodes, replData.MasterIp)
+	case globals.InnoDBClusterLabel:
+		err = CreateInnoDBClusterReplication(sdef, origin, replData.Nodes, replData.MasterIp)
 	}
 	return err
 }